hide typesense behind impl
