+++ config/config.exs
@@ -9,6 +9,7 @@ config :hexdocs,
   hexpm_impl: Hexdocs.Hexpm.Impl,
   store_impl: Hexdocs.Store.Local,
   cdn_impl: Hexdocs.CDN.Local,
+  search_impl: Hexdocs.Search.Local,
   source_repo_impl: Hexdocs.SourceRepo.GitHub,
   tmp_dir: "tmp",
   queue_id: "test",
