+++ config/dev.exs
@@ -5,4 +5,5 @@ config :hexdocs,
   hexpm_url: "http://localhost:4000",
   hexpm_impl: Hexdocs.Hexpm.Impl,
   store_impl: Hexdocs.Store.Local,
-  cdn_impl: Hexdocs.CDN.Local
+  cdn_impl: Hexdocs.CDN.Local,
+  search_impl: Hexdocs.Search.Local
