+++ config/prod.exs
@@ -4,6 +4,7 @@ config :hexdocs,
   hexpm_impl: Hexdocs.Hexpm.Impl,
   store_impl: Hexdocs.Store.Impl,
   cdn_impl: Hexdocs.CDN.Fastly,
+  search_impl: Hexdocs.Search.Typesense,
   queue_producer: BroadwaySQS.Producer,
   gcs_put_debounce: 3000
